Trying to reset baseline... trial and error
After a lot of trial and error and trying to refactor a lot of things
out I am going back to how it was…  Packer over WinRM has lots of
issues.  sometimes in the process of uploading a provisioning script it
errors out, or copying a file.  Instead of using packer to name all the
provisioning scripts and file uploads I am creating 1 provision script
that handles all the provisioning and 1 cleanup script that packer calls